backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python-headless
numpy
pillow
ultralytics